+++ kopy-compiler/test/kotlin/com/javiersc/kotlin/kopy/compiler/KopyCompilerPluginRegistrarTest.kt
@@ -5,10 +5,11 @@ import com.javiersc.kotlin.kopy.compiler.ir.IrKopyGenerationExtension
 import com.javiersc.kotlin.stdlib.endWithNewLine
 import com.javiersc.kotlin.stdlib.removeDuplicateEmptyLines
 import com.javiersc.kotlin.stdlib.second
-import io.kotest.matchers.collections.shouldContain
-import io.kotest.matchers.collections.shouldHaveSize
-import io.kotest.matchers.shouldBe
-import io.kotest.matchers.types.shouldBeTypeOf
+import com.javiersc.kotlin.test.assertContains
+import com.javiersc.kotlin.test.assertCount
+import com.javiersc.kotlin.test.assertEquals
+import com.javiersc.kotlin.test.assertIs
+import com.javiersc.kotlin.test.assertTrue
 import java.io.File
 import kotlin.io.path.createTempDirectory
 import kotlin.time.Duration.Companion.milliseconds
@@ -22,7 +23,7 @@ class KopyCompilerPluginRegistrarTest {
 
     @Test
     fun `check KopyCompilerPluginRegistrar uses K2`() {
-        KopyCompilerPluginRegistrar().supportsK2 shouldBe true
+        KopyCompilerPluginRegistrar().supportsK2.assertTrue()
     }
 
     @Test
@@ -32,9 +33,9 @@ class KopyCompilerPluginRegistrarTest {
         val kopyCompiler = KopyCompilerPluginRegistrar()
         extensionStorage.run { kopyCompiler.run { registerExtensions(configuration) } }
         val extensions: List<Any> = extensionStorage.registeredExtensions.values.flatten()
-        extensions.shouldHaveSize(2)
-        extensions.first().shouldBeTypeOf<FirKopyExtension>()
-        extensions.second().shouldBeTypeOf<IrKopyGenerationExtension>()
+        extensions.assertCount(2)
+        extensions.first().assertIs<FirKopyExtension>()
+        extensions.second().assertIs<IrKopyGenerationExtension>()
     }
 
     @Test
@@ -43,7 +44,7 @@ class KopyCompilerPluginRegistrarTest {
         val configuration = CompilerConfiguration()
         val kopyCompiler = KopyCompilerPluginRegistrar()
         extensionStorage.run { kopyCompiler.run { registerExtensions(configuration) } }
-        extensionStorage.disposables.firstIsInstance<MeasureStorage>().measures.shouldHaveSize(0)
+        extensionStorage.disposables.firstIsInstance<MeasureStorage>().measures.assertCount(0)
     }
 
     @Test
@@ -74,18 +75,19 @@ class KopyCompilerPluginRegistrarTest {
             .replace("\r\n", "\n")
             .removeDuplicateEmptyLines()
             .endWithNewLine()
-            .shouldBe(
-                """ |KEY_A:
-                    |  - VALUE_A, 1s
-                    |  - VALUE_A_A, 500ms
-                    |KEY_B:
-                    |  - VALUE_B, 150ms
-                    |KEY_C:
-                    |  - VALUE_C_C_C, 800us
-                    |  - VALUE_C_C, 600us
-                    |  - VALUE_C, 300us
-                    |
-                    |TOTAL: 1.651700s
+            .assertEquals(
+                """
+                |KEY_A:
+                |  - VALUE_A, 1s
+                |  - VALUE_A_A, 500ms
+                |KEY_B:
+                |  - VALUE_B, 150ms
+                |KEY_C:
+                |  - VALUE_C_C_C, 800us
+                |  - VALUE_C_C, 600us
+                |  - VALUE_C, 300us
+                |
+                |TOTAL: 1.651700s
                 """
                     .trimMargin()
                     .removeDuplicateEmptyLines()
@@ -95,7 +97,7 @@ class KopyCompilerPluginRegistrarTest {
         kopyConfig.measureExecution("KEY_Z") { "VALUE_Z" }
         kopyConfig.measureStorage.measures
             .map { Triple(it.key, it.value, it.isNested) }
-            .shouldContain(Triple("KEY_Y", "VALUE_Y", true))
-            .shouldContain(Triple("KEY_Z", "VALUE_Z", false))
+            .assertContains(Triple("KEY_Y", "VALUE_Y", true))
+            .assertContains(Triple("KEY_Z", "VALUE_Z", false))
     }
 }